requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
duckduckgo-search
arxiv